# connectors/docs/swagger.yaml
definitions:
...
  fiber.Map:
    additionalProperties: {}
    type: object
...
  irminmodels.AlreadyRunningBody:
    properties:
      error:
        description: |-
          Error is the stable human-readable message. Always
          "Operation is already running" — kept for backwards
          compatibility with clients that key off the error string.
        example: Operation is already running
        type: string
      job_id:
        description: |-
          JobID is the opaque identifier of the currently-running job.
          Clients can pass this straight to CancelOperationJob or
          GetOperationJobStatus. Empty only for legacy holders that have
          not migrated to the JobManager.Begin path yet.
        example: opjob_9m3x7k2n8q5p
        type: string
      kind:
        description: |-
          Kind is the operation kind holding the lock. One of "pull",
          "push", "patch", "schema". Helps operators distinguish a
          long-running pull from a fast sync push when triaging
          contention.
        example: pull
        type: string
      operation_id:
        description: |-
          OperationID is the connector-service-local numeric ID of the
          operation the lock was taken on. Present so operator-facing
          logs can cross-reference the operation row directly.
        example: 42
        type: integer
      started_at:
        description: |-
          StartedAt is when the in-flight job was accepted. Clients can
          compute elapsed time to decide whether to wait or cancel. nil
          for legacy sync holders with no registered row.

          Pointer rather than time.Time + omitempty because encoding/json's
          omitempty does not treat a zero time.Time as empty — it would
          serialise as "0001-01-01T00:00:00Z" on the wire. Matches the
          convention used by ExpiresAt above and the *time.Time fields
          in the other models in this package.
        type: string
    type: object
...
  irminmodels.JobErrorBody:
    properties:
      error:
        description: |-
          Error is the operator-facing human-readable message. Stable
          enough for log grep but not a machine-readable code.
        example: failed to load job
        type: string
      job_id:
        description: |-
          JobID echoes the ID from the URL when the server could parse
          one. Empty for request-level failures (malformed URL, missing
          path parameter).
        example: opjob_9m3x7k2n8q5p
        type: string
      reason:
        allOf:
        - $ref: '#/definitions/irminmodels.JobErrorReason'
        description: |-
          Reason is the machine-readable failure classification. See
          JobErrorReason for the stable value set.
        example: transient_db_error
      retryable:
        description: |-
          Retryable reports whether the caller should retry the same
          request after a short backoff. False for any terminal failure
          mode (not_found, corrupted_job_state, invalid_request).
        example: true
        type: boolean
    type: object
  irminmodels.JobErrorReason:
    enum:
    - transient_db_error
    - corrupted_job_state
    - not_found
    - invalid_request
    - internal
    type: string
    x-enum-varnames:
    - JobErrorReasonTransientDB
    - JobErrorReasonCorruptedRow
    - JobErrorReasonNotFound
    - JobErrorReasonInvalidRequest
    - JobErrorReasonInternal
...
  irminmodels.StartOperationJobResponse:
    properties:
      job_id:
        description: |-
          JobID is the identifier the caller uses on subsequent
          /operation/status and /operation/result calls.
        example: opjob_9m3x7k2n8q5p
        type: string
      operation_token:
        description: |-
          OperationToken is the per-job bearer credential for lifecycle
          routes. Minted by the server on Start*; TTL matches the
          underlying OperationJob row.
        example: optk_7f3d2a9c1e6b
        type: string
    type: object
...
host: connectors.irmin.co
info:
  contact:
    email: hello@irmin.co
    name: Irmin Team
    url: https://irmin.co
  description: Collection of connectors for Irmin to interact with external services
  termsOfService: http://irmin.co/terms
  title: Irmin Connectors
  version: 1.0.0
paths:
...
  /firecrawl/operation/patch:
    post:
      consumes:
      - multipart/form-data
      description: Firecrawl connector does not support patch operations - it is read-only
      produces:
      - application/json
      responses:
        "401":
          description: Unauthorized - invalid or missing authentication
          schema:
            $ref: '#/definitions/fiber.Map'
        "501":
          description: Not implemented - Firecrawl connector does not support patch
            operations
          schema:
            $ref: '#/definitions/fiber.Map'
      security:
      - SystemTokenAuth: []
      summary: Firecrawl connector does not support patch operations
      tags:
      - firecrawl
...
  /googledrive/operation/patch:
    post:
      consumes:
      - multipart/form-data
      description: Google Drive connector does not support patch operations.
      produces:
      - application/json
      responses:
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/fiber.Map'
        "501":
          description: Not implemented - Google Drive does not support patch
          schema:
            $ref: '#/definitions/fiber.Map'
      security:
      - SystemTokenAuth: []
      summary: Google Drive connector does not support patch operations
      tags:
      - googledrive
...
  /http/operation/patch:
    post:
      consumes:
      - multipart/form-data
      description: Apply JSON Patch operations by making HTTP requests to the configured
        endpoint. Each patch operation becomes an HTTP request.
      parameters:
      - description: JSON file containing array of patch operations
        in: formData
        name: patches
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: Patch operations applied successfully
          schema:
            $ref: '#/definitions/fiber.Map'
        "400":
          description: Bad request - invalid patch format
          schema:
            $ref: '#/definitions/fiber.Map'
        "401":
          description: Unauthorized - invalid or missing authentication
          schema:
            $ref: '#/definitions/fiber.Map'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/fiber.Map'
      security:
      - SystemTokenAuth: []
      summary: Apply patch operations via HTTP requests
      tags:
      - http
...
  /linear/operation/patch:
    post:
      consumes:
      - multipart/form-data
      description: Translates a JSON-Patch document into one save_issue MCP tool call
        per targeted Linear issue. Only add/replace/remove are supported. The patch
        path must reference an issue file (e.g., `/issues/<uuid>.json/title`) where
        `<uuid>` is the Linear UUID, not the identifier (IRM-42).
      parameters:
      - description: JSON file containing a JSON-Patch array
        in: formData
        name: patches
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: Patches applied successfully
          schema:
            $ref: '#/definitions/fiber.Map'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/fiber.Map'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/fiber.Map'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/fiber.Map'
      security:
      - SystemTokenAuth: []
      summary: Apply JSON-Patch operations to Linear issues
      tags:
      - linear
...
  /mysql/operation/patch:
    post:
      consumes:
      - multipart/form-data
      description: Apply granular updates to MySQL database records using JSON patch
        operations
      parameters:
      - description: JSON patch file containing update operations
        in: formData
        name: patch
        required: true
        type: file
      produces:
      - application/json
      responses:
        "202":
          description: Patch job accepted; poll /operation/status/:job_id
          schema:
            $ref: '#/definitions/irminmodels.StartOperationJobResponse'
        "400":
          description: Bad request - invalid operation token or patch format
          schema:
            $ref: '#/definitions/fiber.Map'
        "401":
          description: Unauthorized - invalid or missing authentication
          schema:
            $ref: '#/definitions/fiber.Map'
        "404":
          description: Operation not found
          schema:
            $ref: '#/definitions/fiber.Map'
        "409":
          description: Operation already running
          schema:
            $ref: '#/definitions/irminmodels.AlreadyRunningBody'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/irminmodels.JobErrorBody'
      security:
      - SystemTokenAuth: []
      summary: Patch data in MySQL database
      tags:
      - mysql
...
  /pinecone/operation/patch:
    post:
      consumes:
      - multipart/form-data
      description: Apply JSON Patch operations to vectors in the Pinecone index. Supports
        add (upsert), remove (delete), and replace (upsert) operations.
      parameters:
      - description: JSON file containing array of patch operations
        in: formData
        name: patches
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: Patch operations applied successfully
          schema:
            $ref: '#/definitions/fiber.Map'
        "400":
          description: Bad request - invalid patch format
          schema:
            $ref: '#/definitions/fiber.Map'
        "401":
          description: Unauthorized - invalid or missing authentication
          schema:
            $ref: '#/definitions/fiber.Map'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/fiber.Map'
      security:
      - SystemTokenAuth: []
      summary: Apply patch operations to Pinecone vectors
      tags:
      - pinecone
...
  /postgres/operation/patch:
    post:
      consumes:
      - multipart/form-data
      description: Apply granular updates to PostgreSQL database records using JSON
        patch operations
      parameters:
      - description: JSON patch file containing update operations
        in: formData
        name: patch
        required: true
        type: file
      produces:
      - application/json
      responses:
        "202":
          description: Patch job accepted; poll /operation/status/:job_id
          schema:
            $ref: '#/definitions/irminmodels.StartOperationJobResponse'
        "400":
          description: Bad request - invalid operation token or patch format
          schema:
            $ref: '#/definitions/fiber.Map'
        "401":
          description: Unauthorized - invalid or missing authentication
          schema:
            $ref: '#/definitions/fiber.Map'
        "404":
          description: Operation not found
          schema:
            $ref: '#/definitions/fiber.Map'
        "409":
          description: Operation already running
          schema:
            $ref: '#/definitions/irminmodels.AlreadyRunningBody'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/irminmodels.JobErrorBody'
      security:
      - SystemTokenAuth: []
      summary: Patch data in PostgreSQL database
      tags:
      - postgres
...
  /sftp/operation/patch:
    post:
      consumes:
      - multipart/form-data
      description: Apply JSON Patch operations to files on the SFTP server. Supports
        add (upload), remove (delete), replace (overwrite), move, and copy operations.
      parameters:
      - description: JSON file containing array of patch operations
        in: formData
        name: patches
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: Patch operations applied successfully
          schema:
            $ref: '#/definitions/fiber.Map'
        "400":
          description: Bad request - invalid patch format
          schema:
            $ref: '#/definitions/fiber.Map'
        "401":
          description: Unauthorized - invalid or missing authentication
          schema:
            $ref: '#/definitions/fiber.Map'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/fiber.Map'
      security:
      - SystemTokenAuth: []
      summary: Apply patch operations to SFTP files
      tags:
      - sftp
...
  /stripe/operation/patch:
    post:
      consumes:
      - multipart/form-data
      description: Translate a JSON-Patch document into one partial-update call per
        targeted Stripe resource. Only add / replace / remove are supported; move
        and copy are rejected because Stripe's fields aren't structurally rearrangeable.
      parameters:
      - description: JSON file containing a JSON-Patch array
        in: formData
        name: patches
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: Patches applied successfully
          schema:
            $ref: '#/definitions/fiber.Map'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/fiber.Map'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/fiber.Map'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/fiber.Map'
      security:
      - SystemTokenAuth: []
      summary: Apply JSON-Patch operations to Stripe resources
      tags:
      - stripe
...
schemes:
- https
securityDefinitions:
  OperationTokenAuth:
    description: Operation token with Bearer prefix - received from operation/init
      and used for data operation endpoints
    in: header
    name: Authorization
    type: apiKey
  SystemTokenAuth:
    description: System token with Bearer prefix - used for connector registration
      and initialization endpoints
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"